backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
docusign-esign==3.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
requests==2.28.2
python-dateutil==2.8.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.11
alembic>=1.8.0
strawberry-graphql[fastapi]>=0.188.0
python-jose[cryptography]==3.3.0
authlib==1.2.1
itsdangerous==2.1.2
python-multipart==0.0.6
email-validator==2.0.0.post2
python-jose[pycryptodome]==3.3.0
